samples/apps/autogen-studio/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "autogenstudio"
authors = [
    { name = "AutoGen Team", email = "autogen@microsoft.com" },
]
description = "AutoGen Studio"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9, <3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]


dependencies = [
    "pyjwt",
    "pymongo",
    "pydantic",
    "fastapi",
    "typer",
    "uvicorn",
    "arxiv",
    "pyautogen[gemini]>=0.2.0",
    "python-dotenv",
    "websockets",
    "numpy < 2.0.0",
    "sqlmodel",
    "psycopg",
    "alembic",
    "loguru",
]
optional-dependencies = { web = ["fastapi", "uvicorn"], database = ["psycopg"] }

dynamic = ["version"]

[tool.setuptools]
include-package-data = true


[tool.setuptools.dynamic]
version = { attr = "autogenstudio.version.VERSION" }
readme = { file = ["README.md"] }

[tool.setuptools.packages.find]
include = ["autogenstudio*"]
exclude = ["*.tests*"]
namespaces = false

[tool.setuptools.package-data]
"autogenstudio" = ["*.*"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Deprecated call to `pkg_resources\\.declare_namespace\\('.*'\\):DeprecationWarning",
    "ignore::DeprecationWarning:google.rpc",
]


[project.urls]
"Homepage" = "https://github.com/microsoft/autogen"
"Bug Tracker" = "https://github.com/microsoft/autogen/issues"

[project.scripts]
autogenstudio = "autogenstudio.cli:run"
